inference_service/requirements.txt
pycoral
pyzmq
pillow
numpy
opencv-python-headless
